apps/backend/requirements.txt
# apps/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.0
pandas==2.2.3
numpy==2.2.4
requests==2.32.3
plotly==5.24.1
supabase==2.15.0
httpx==0.28.1
huggingface_hub==0.30.2
duckdb==1.2.2
polars==1.28.0
pydantic-ai==0.3.0
pydantic>=2.6.3
logfire>=0.30.0
devtools>=0.12.2
asyncpg>=0.29.0
annotated_types>=0.6.0
typing_extensions>=4.10.0
drizzle-orm==2.0.1






